requirements.txt
# Core dependencies for HELD product vectorization workflow
openai>=1.0.0
qdrant-client>=1.7.0
numpy>=1.24.0

# Optional dependencies for enhanced functionality
python-dotenv>=1.0.0
requests>=2.31.0

# Database dependencies
psycopg2-binary>=2.9.0